refactor(ui): split navigation hub components
